Change pdf.format back to pdf.pattern; plus, 1120S
Both the code for Form 1120S and the code for its Schedule K have now
been rewritten for the new PDF class, whose behavior should now be
stable; having tried several experiments to complexify it, I have
decided that the pdf.pattern trick is enough, and have eliminated the
idea of auto-wrapping or auto-formatting values.  Zen of Python ftw.