Remove a date restriction specific to my ledger
My main ledger starts on 2012/1/1 and during that first January includes
a stray 941 payment that has nothing to do with the 941 calculations in
the rest of the file, since they all have to do with 2012 obligations,
whereas that first payment is discharging the rest of my 2011
obligation.  As you can see, I used to protect against that by having a
restriction built into this general forms.dat file.  Now, I instead am
going to limit the range of dates that I process over my own ledger.dat
whenever I use it in conjunction with this forms.dat file, so that this
file can contain only the truth about taxes, without odd corner cases
that only work with my main business ledger and would probably break all
other ledgers I might create in the future that cover the same timespan.