Correct unfortunate variable names in Schedule D
Part II of the schedule does not have a "line 1" and "line 2", but has
lines 3 and 4 of the form.  But using those names would break the
symmetry between the first and second halves of the form, so I am going
with "table" and "total" as the generic names for the upper and lower
parts of both of those pages of the form.