Add form_version as a "defaults" command-line arg
Before, running "defaults" on a form simply filled in a version like
"2012", but in fact that is not adequate since the very defaults
themselves will differ from year to year as fields are added and removed
and renamed.  So each form now has to advertise the versions that it
supports, and the version needs to be provided to the "defaults"
command; if omitted, a list of the available versions is printed.

Also, renamed the awkward "form.form" JSON value to "form.form_name".