Switch to an official Transaction model
Since it is each importer class that should be in charge of the data it
is scanning, and whether that data needs extra fields for metadata
particular to its flavor of bank statement.  Also: used this opportunity
to rough out some REs that signal the beginning and ending of a
particular account in DCCU statements.