Disable auto-choosing form versions
Previously, a form with only one version did not require a second
argument when printed with the "form" sub-command; but this winds up
hiding important information from the user: whether the form supports
the particular year or period that the user needs.  So now, even a form
with only one version will still prompt the user to specify that version.