Revert PyPDF2 work
Given that PyPDF2 is currently unmaintained and that, while I found
patches for everything else in issues and on Stack Overflow, I can’t
find user-supplied code for flattening a PDF, I’ve instead figured out
how to install pdftk on Ubuntu 18.04 and will keep using it for now.
(The answer was a shell script from Stack Overflow.)

I can always un-revert this in the future if PyPDF2 is re-awakened and
learns how to flatten PDF files.